refactor(course-cover): extract shared cover/glyph helpers
CourseCard and BookmarkCourseList carried independent copies of the
COVER_VARIANTS pool plus coverFor / glyphFor. Pull them into a private
module under (protected)/_components so adding a new variant only needs
one edit.
